moved trt7 ops to converters dir
